Fix utilization computation error
int is not enough for cycles. Also change active and total cycles to double to
avoid overflow.

+++ thread.c
@@ -394,15 +394,15 @@ static void *worker_libevent(void *arg) {
     // event_base_loop(me->base, 0);
     uint64_t cycles_per_sec = 2000000000; // 2.0GHz
     uint64_t prev = rdtsc();
-    uint64_t active_cycles = 0;
-    uint64_t total_cycles = 0;
+    double active_cycles = 0.0;
+    double total_cycles = 0.0;
     uint64_t curr = prev;
     uint64_t print_prev = prev;
     while (1) {
         handled_event = false;
         ret = event_base_loop(me->base, EVLOOP_NONBLOCK);
         curr = rdtsc();
-        int delta = curr - prev;
+        uint64_t delta = curr - prev;
         if (handled_event) {
             active_cycles += delta;
         } else if (ret == -1) {
@@ -411,10 +411,10 @@ static void *worker_libevent(void *arg) {
         total_cycles += delta;
         if (curr - print_prev >= cycles_per_sec) {
 #if defined(TIMETRACE) || defined(TIMETRACE_HANDLE)
-            fprintf(stderr, "Utilization is: %.4lf %% \n", (double)active_cycles / total_cycles * 100.0);
+            fprintf(stderr, "Utilization is: %.4lf %% \n", active_cycles / total_cycles * 100.0);
 #endif
-            total_cycles = 0;
-            active_cycles = 0;
+            total_cycles = 0.0;
+            active_cycles = 0.0;
             print_prev = curr;
         }
         prev = curr;
